requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
cattrs==22.2.0
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==39.0.2
daphne==4.0.0
decorator==5.1.1
Django==4.1.7
django-taggit==3.1.0
djangorestframework==3.14.0
docstring-to-markdown==0.11
executing==1.2.0
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipython==8.11.0
jedi==0.18.2
jedi-language-server==0.40.0
lsprotocol==2022.0.0a10
Markdown==3.4.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
promise==2.3
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.5
pygls==1.0.1
Pygments==2.14.0
pyOpenSSL==23.0.0
pytz==2022.7.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
text-unidecode==1.3
traitlets==5.9.0
Twisted==22.10.0
txaio==23.1.1
typeguard==2.13.3
typing_extensions==4.5.0
wcwidth==0.2.6
zope.interface==6.0
